refactor poll store and move to providers/stores
