+++ apps/web/core/test/core.test.ts
@@ -7,7 +7,7 @@ import { PollManagerController } from "../controllers/poll-manager-controller";
 import { WalletController } from "../controllers/wallet-controller";
 import { WalletProvider } from "../providers/wallets/wallet-provider";
 import { PrivateKey } from "o1js";
-import { InMemoryPollStore } from "../controllers/poll-store";
+import { InMemoryPollStore } from "../providers/stores/in-memory-poll-store";
 import { ChainTestController } from "./test-utils/chain-test-controller";
 import { PollController } from "../controllers/poll-controller";
